fix: Food DB의 food_count type 변경 (#16)

+++ backend/src/main/java/org/example/backend/api/myfridge/model/dto/FoodDetailDto.java
@@ -18,7 +18,7 @@ public class FoodDetailDto {
     private Long foodId;
     private String foodListName;
     private LocalDate foodRegistDate;
-    private int foodCount;
+    private float foodCount;
     private FoodUnit foodUnit;
     private LocalDate foodProDate;
     private LocalDate foodExpDate;
