+++ backend/src/main/java/org/example/backend/api/myfridge/model/entity/Food.java
@@ -49,7 +49,7 @@ public class Food {
   private LocalDate foodRegistDate;
 
   @NotNull
-  private int foodCount;
+  private float foodCount;
 
   @Enumerated(EnumType.STRING)
   @NotNull
